requirements.txt
streamlit
google-genai
python-dotenv
pdf2image
python-docx
pdfplumber
PyPDF2
pillow
sentence-transformers
faiss-cpu
